UI and styling overhaul for enhanced UX
Significant updates were made to both the user interface and the styling of the application to improve the user experience. In the AuthenticationLayout.razor and Home.razor files, content was updated to include placeholder text for ongoing development and a more personalized introduction for new users, respectively. Conditional rendering logic was refined to better distinguish between new and logged-in users, alongside the addition of a FriendsList component for all users.

CSS changes encompassed a wide range of adjustments for better layout control, responsiveness, and visual appeal. This includes the introduction of specific margin and padding classes for precise spacing, enhancements in grid and flexbox utilities for improved layout flexibility, and updates to utility classes for sizing, display, and responsive design. Additionally, the color scheme, typography, and border radius adjustments reflect a modernized design aesthetic. Miscellaneous changes to overflow properties and focus/hover states were also made to enhance accessibility and interactivity.

Overall, these changes aim to provide a more engaging and accessible user interface while laying the groundwork for further development and testing.